+++ server/src/main/java/org/apache/iotdb/db/metadata/rescon/SchemaResourceManager.java
@@ -21,8 +21,7 @@ package org.apache.iotdb.db.metadata.rescon;
 
 import org.apache.iotdb.commons.service.metric.MetricService;
 import org.apache.iotdb.db.conf.IoTDBDescriptor;
-import org.apache.iotdb.db.metadata.mtree.store.disk.MTreeFlushTaskManager;
-import org.apache.iotdb.db.metadata.mtree.store.disk.MTreeReleaseTaskManager;
+import org.apache.iotdb.db.metadata.mtree.store.disk.cache.CacheMemoryManager;
 import org.apache.iotdb.db.metadata.mtree.store.disk.memcontrol.MemManagerHolder;
 import org.apache.iotdb.db.metadata.schemaregion.SchemaEngineMode;
 
@@ -58,14 +57,11 @@ public class SchemaResourceManager {
   private static void initSchemaFileModeResource() {
     MemManagerHolder.initMemManagerInstance();
     MemManagerHolder.getMemManagerInstance().init();
-    MTreeFlushTaskManager.getInstance().init();
-    MTreeReleaseTaskManager.getInstance().init();
+    CacheMemoryManager.getInstance().init();
   }
 
   private static void clearSchemaFileModeResource() {
     MemManagerHolder.getMemManagerInstance().clear();
-    // the release task may submit flush task, thus must be shut down and clear first
-    MTreeReleaseTaskManager.getInstance().clear();
-    MTreeFlushTaskManager.getInstance().clear();
+    CacheMemoryManager.getInstance().clear();
   }
 }
